requirements.txt
boto3==1.9.194
psycopg2-binary==2.8.3
